# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ParityLab API
  version: 0.1.0
  description: Deterministic, sandbox-only Stripe integration verification.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/settings/project:
    get:
      operationId: getProjectSettings
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Current project settings.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectSettings' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateProjectSettings
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                retention_days: { type: integer, minimum: 1, maximum: 3650 }
      responses:
        '200':
          description: Persisted project settings.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /v1/environments/{id}/select:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: selectEnvironment
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Newly selected project environment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/findings/{id}/resolve:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: resolveFinding
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Persisted resolved finding.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Finding' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/findings/{id}/reopen:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: reopenFinding
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Persisted reopened finding.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Finding' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Count of unread notifications marked read.
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/notifications/{id}/read:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: markNotificationRead
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Persisted read notification.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: paritylab_session
      description: Opaque session token. Only its SHA-256 digest is persisted server-side.
  parameters:
...
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Stripe-style API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ProjectSettings:
      type: object
      additionalProperties: false
      required: [id, name, retention_days]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        retention_days: { type: integer, minimum: 1, maximum: 3650 }
    Environment:
      type: object
      additionalProperties: false
      required: [id, name, kind, is_default]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [local, sandbox, staging] }
        is_default: { type: boolean }
    Notification:
      type: object
      additionalProperties: false
      required: [id, kind, payload, created_at]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string }
        kind: { type: string }
        payload: { type: object, additionalProperties: true }
        read_at:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        created_at: { type: string, format: date-time }
...
    Finding:
      type: object
      required: [id, severity, title, summary, cause, remediation, checkpoint, resolved]
      properties:
        id: { type: string }
        severity: { type: string, enum: [info, warning, critical] }
        title: { type: string }
        summary: { type: string }
        cause: { type: string }
        remediation: { type: string }
        checkpoint: { type: string }
        resolved: { type: boolean }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, code, message, request_id]
          properties:
            type: { type: string }
            code: { type: string }
            message: { type: string }
            param: { type: string }
            request_id: { type: string }